Give tomes custom tags when crafting
There's not much documentation on how to give ItemStacks custom
NBTs at init. This covers the case when a tome is crafted, but
will not assign tags if the tome is pulled in from the Creative
inventory or given using the '/give' command (unless the tags are
specified in the command). This results in tomes given by commands
or pulled from the Creative inventory not stacking with other tomes
until they are used (the 'use' method will write the tags if they
are missing).