Change point transfers to not use a curve
Previously, tomes relied on knowledge of the experience point curve
<https://minecraft.fandom.com/wiki/Experience> to figure out how many
points were needed to advance or go back a number of levels.

Now, tomes use only knowledge of the current experience level as
provided by the PlayerEnity class. This should make the mod more
compatible with other mods that provide tweaks to the experience curve.